Rename handler module
