+++ backend/app/services/dependencies.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from fastapi import Depends
 
-from .chat_handler import ChatHandler
+from .relation_handler import ChatHandler
 from .interfaces import AbstractChatHandler, AbstractRecordHandler
 from .record_handler import RecordHandler
